+++ R/sgo_transform.R
@@ -16,7 +16,6 @@ NULL
 #' @param ... Additional parameters passed to internal functions. Currently it
 #' supports the additional arguments seen in \code{sgo_bng_lonlat} and
 #' \code{sgo_lonlat_bng}.
-#' this description and which parameters may admit...
 #' @details
 #' This function is a wrapper of specific transformation functions
 #' (\code{\link{sgo_bng_lonlat}}, \code{\link{sgo_en_wgs84}},
@@ -365,7 +364,7 @@ sgo_transform.sgo_points <- function(x, to=NULL, ...) {
 .FUN_TO_4936 <- list(list(sgo_set_gcs, sgo_lonlat_cart),
                      list(sgo_en_wgs84, sgo_set_gcs, sgo_lonlat_cart),     #3857
                      list(sgo_lonlat_bng, sgo_bng_lonlat, sgo_lonlat_cart),#4277
-                     list(sgo_bng_lonlat, sgo_lonlat_cart),
+                     list(sgo_bng_lonlat, sgo_lonlat_cart),               #27700
                      list(sgo_lonlat_cart),                                #4258
                      list(sgo_set_gcs, sgo_lonlat_cart),                   #4979
                      list(sgo_cart_lonlat, sgo_set_gcs, sgo_lonlat_cart),  #4978
@@ -377,7 +376,7 @@ sgo_transform.sgo_points <- function(x, to=NULL, ...) {
 .ARGS_TO_4936 <- list(list(list(to=4258), list()),
                       list(list(to=4326), list(to=4258), list()),
                       list(list(to=27700), list(to=4258), list()),
-                      list(list(to=4258), list()),
+                      list(list(to=4258), list()),                        #27700
                       list(),                                              #4258
                       list(list(to=4937), list()),
                       list(list(), list(to=4937), list()),