+++ data-raw/build_sysdata.R
@@ -110,11 +110,12 @@ OSTN15 <- read.csv(file)
                                      "ETRS89_ODN_HeightShift",
                                      "Height_Datum_Flag")])
 colnames(.ostn.shifts) <- c("e", "n", "g", "f")
+.ROWS.OSTN.SHIFTS <- nrow(.ostn.shifts)
 
 
 #SAVE EVERYTHING IN sysdata.rda
 save(RAD.TO.GRAD, PI, .epsgs, .sgo_points.attr, .sgo_points.2d.coords,
      .sgo_points.3d.coords, .sgo_points.2d.core, .sgo_points.3d.core,
      .coordinates.names, lonlat.ellipsoid, lonlat.datum, datum.flags,
-     .ostn.shifts, file = "./R/sysdata.rda", compress = "xz")
+     .ostn.shifts, .ROWS.OSTN.SHIFTS, file = "./R/sysdata.rda", compress = "xz")
 #load(file = "./R/sysdata.rda")