+++ NEWS.md
@@ -1,3 +1,10 @@
+# sgo 0.9.0.9000
+
+* Improved some parts of sgo_transforms's documentation.
+* Fixed the output text of the print function.
+* Fixed: Points right on the upper boundaries of the OSTN grid are considered   
+  outside of the OSTN15 rectangle
+
 # sgo 0.9.0
 
 * Initial release on CRAN.